feat: unpack CardOps in ResCardOps and DevCardOps
