+++ src/main/scala/scatan/model/game/ScatanEffects.scala
@@ -4,14 +4,10 @@ import scatan.lib.game.ops.Effect
 import scatan.model.components.{BuildingType, ResourceCard}
 import scatan.model.game.config.ScatanActions.*
 import scatan.model.game.config.ScatanPlayer
+import scatan.model.game.ops.ResCardOps.{assignResourcesFromNumber, stoleResourceCard}
+import scatan.model.game.ops.DevCardOps.buyDevelopmentCard
+
 import scatan.model.game.ops.BuildingOps.{assignBuilding, build}
-import scatan.model.game.ops.CardOps.{
-  assignResourceCard,
-  assignResourcesFromNumber,
-  buyDevelopmentCard,
-  removeResourceCard,
-  stoleResourceCard
-}
 import scatan.model.map.{Hexagon, RoadSpot, StructureSpot}
 import scatan.model.components.ResourceCard
 import scatan.model.game.ops.TradeOps.tradeWithPlayer
