+++ src/test/scala/scatan/model/game/ops/AwardOpsTest.scala
@@ -4,8 +4,8 @@ import scatan.model.components.*
 import scatan.model.game.{BaseScatanStateTest, ScatanState}
 import scatan.model.game.ops.AwardOps.*
 import scatan.model.game.ops.BuildingOps.assignBuilding
-import scatan.model.game.ops.CardOps.assignDevelopmentCard
-import scatan.model.game.ops.EmptySpotsOps.{emptyRoadSpot, emptyStructureSpot}
+import scatan.model.game.ops.DevCardOps.assignDevelopmentCard
+import scatan.model.game.ops.EmptySpotOps.{emptyRoadSpot, emptyStructureSpot}
 
 class AwardOpsTest extends BaseScatanStateTest:
