+++ src/test/scala/scatan/model/game/ops/BuildingOpsTest.scala
@@ -4,8 +4,8 @@ import scatan.model.components.*
 import scatan.model.game.ScatanState
 import scatan.model.game.{BaseScatanStateTest, ScatanState}
 import scatan.model.game.ops.BuildingOps.{assignBuilding, build}
-import scatan.model.game.ops.CardOps.assignResourceCard
-import scatan.model.game.ops.EmptySpotsOps.{emptyStructureSpot, emptyRoadSpot}
+import scatan.model.game.ops.ResCardOps.assignResourceCard
+import scatan.model.game.ops.EmptySpotOps.{emptyStructureSpot, emptyRoadSpot}
 import scatan.model.game.BaseScatanStateTest
 import scatan.model.game.config.ScatanPlayer
 import scatan.model.map.{RoadSpot, StructureSpot}
