+++ src/test/scala/scatan/model/game/ops/DevCardOpsTest.scala
@@ -3,12 +3,8 @@ package scatan.model.game.ops
 import scatan.lib.game.Game
 import scatan.model.components.{DevelopmentCard, DevelopmentType, ResourceCard, ResourceType}
 import scatan.model.game.{BaseScatanStateTest, ScatanState}
-import scatan.model.game.ops.CardOps.{
-  assignDevelopmentCard,
-  assignResourceCard,
-  buyDevelopmentCard,
-  consumeDevelopmentCard
-}
+import scatan.model.game.ops.DevCardOps.{buyDevelopmentCard, consumeDevelopmentCard, assignDevelopmentCard}
+import scatan.model.game.ops.ResCardOps.assignResourceCard
 
 class DevCardOpsTest extends BaseScatanStateTest:
