+++ src/test/scala/scatan/model/game/ops/ScoreOpsTest.scala
@@ -3,9 +3,9 @@ package scatan.model.game.ops
 import scatan.model.components.{AssignedBuildings, BuildingType, Scores}
 import scatan.model.game.{BaseScatanStateTest, ScatanState}
 import scatan.model.game.ops.BuildingOps.assignBuilding
-import scatan.model.game.ops.EmptySpotsOps.{emptyRoadSpot, emptyStructureSpot}
+import scatan.model.game.ops.EmptySpotOps.{emptyRoadSpot, emptyStructureSpot}
 import scatan.model.game.ops.ScoreOps.*
-import scatan.model.game.ops.CardOps.assignDevelopmentCard
+import scatan.model.game.ops.DevCardOps.assignDevelopmentCard
 import scatan.model.components.DevelopmentCard
 import scatan.model.components.DevelopmentType
